Move files to match common structure
