+++ src/components/charts/RowChart/RowChartSettingsPanel.tsx
